requirements.txt
Django==3.1.6
djangorestframework==3.12.2
gunicorn==20.0.4
Pillow==8.1.0
psycopg2-binary==2.8.6
pyTelegramBotAPI==3.7.6
requests==2.26.0
